Move pythogoras document fetching/caching to own classes
