+++ src/ume-app-estateservice/Umea.se.EstateService.Logic/DependencyInjectorLogic.cs
@@ -1,6 +1,7 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection;
 using Umea.se.EstateService.Logic.Handlers;
 using Umea.se.EstateService.Logic.Interfaces;
+using Umea.se.EstateService.Logic.Providers;
 
 namespace Umea.se.EstateService.Logic;
 
@@ -9,6 +10,10 @@ public static class DependencyInjectorLogic
     public static IServiceCollection AddLogicDependencies(this IServiceCollection services)
     {
         services.AddSingleton<IPythagorasHandler, PythagorasHandler>();
+        services.AddSingleton<SearchHandler>();
+        services.AddSingleton<PythagorasDocumentProvider>();
+        services.AddSingleton<IPythagorasDocumentProvider, CachedPythagorasDocumentProvider>();
+
         return services;
     }
 }
