+++ src/ume-app-estateservice/Umea.se.EstateService.Logic/Interfaces/IPythagorasDocumentProvider.cs
@@ -1,0 +1,8 @@
+using Umea.se.EstateService.Shared.Search;
+
+namespace Umea.se.EstateService.Logic.Interfaces;
+
+public interface IPythagorasDocumentProvider
+{
+    Task<ICollection<PythagorasDocument>> GetDocumentsAsync();
+}
